tests(e2e): move setup and teardown functions
